// Swagger/models/LoginResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Shift
 * Shift Server API documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LoginResponse
 */
export interface LoginResponse {
    /**
     * 
     * @type {string}
     * @memberof LoginResponse
     */
    accessToken?: string;
    /**
     * 
     * @type {string}
     * @memberof LoginResponse
     */
    msg?: string | null;
    /**
     * 
     * @type {string}
     * @memberof LoginResponse
     */
    passwordMessage?: string;
    /**
     * 
     * @type {string}
     * @memberof LoginResponse
     */
    usernameMessage?: string;
}

export function LoginResponseFromJSON(json: any): LoginResponse {
    return LoginResponseFromJSONTyped(json, false);
}

export function LoginResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): LoginResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'accessToken': !exists(json, 'accessToken') ? undefined : json['accessToken'],
        'msg': !exists(json, 'msg') ? undefined : json['msg'],
        'passwordMessage': !exists(json, 'passwordMessage') ? undefined : json['passwordMessage'],
        'usernameMessage': !exists(json, 'usernameMessage') ? undefined : json['usernameMessage'],
    };
}

export function LoginResponseToJSON(value?: LoginResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'accessToken': value.accessToken,
        'msg': value.msg,
        'passwordMessage': value.passwordMessage,
        'usernameMessage': value.usernameMessage,
    };
}